拆分 DownloaderManager
