+++ src/main/java/com/ghostchu/peerbanhelper/module/impl/webapi/PBHOOBEController.java
@@ -2,6 +2,7 @@ package com.ghostchu.peerbanhelper.module.impl.webapi;
 
 import com.ghostchu.peerbanhelper.Main;
 import com.ghostchu.peerbanhelper.downloader.Downloader;
+import com.ghostchu.peerbanhelper.downloader.DownloaderManager;
 import com.ghostchu.peerbanhelper.module.AbstractFeatureModule;
 import com.ghostchu.peerbanhelper.text.Lang;
 import com.ghostchu.peerbanhelper.util.context.IgnoreScan;
@@ -31,6 +32,8 @@ import static com.ghostchu.peerbanhelper.text.TextManager.tl;
 public final class PBHOOBEController extends AbstractFeatureModule {
     @Autowired
     private JavalinWebContainer webContainer;
+    @Autowired
+    private DownloaderManager downloaderManager;
 
     @Override
     public boolean isConfigurable() {
@@ -75,8 +78,8 @@ public final class PBHOOBEController extends AbstractFeatureModule {
             throw new IllegalArgumentException("Illegal character (.) in name: " + name);
         }
         JsonObject config = draftDownloader.get("config").getAsJsonObject();
-        Downloader downloader = getServer().createDownloader(name, config);
-        if (getServer().registerDownloader(downloader)) {
+        Downloader downloader = downloaderManager.createDownloader(name, config);
+        if (downloaderManager.registerDownloader(downloader)) {
             ctx.status(HttpStatus.CREATED);
             ctx.json(new StdResp(true, tl(locale(ctx), Lang.DOWNLOADER_API_CREATED), null));
         } else {
@@ -84,7 +87,7 @@ public final class PBHOOBEController extends AbstractFeatureModule {
             ctx.json(new StdResp(false, tl(locale(ctx), Lang.DOWNLOADER_API_CREATION_FAILED_ALREADY_EXISTS), null));
         }
         try {
-            getServer().saveDownloaders();
+            downloaderManager.saveDownloaders();
         } catch (IOException e) {
             log.error("Internal server error, unable to create downloader due an I/O exception", e);
             ctx.status(HttpStatus.INTERNAL_SERVER_ERROR);
@@ -109,7 +112,7 @@ public final class PBHOOBEController extends AbstractFeatureModule {
 //            ctx.json(Map.of("message", Lang.DOWNLOADER_API_TEST_NAME_EXISTS));
 //            return;
 //        }
-        Downloader downloader = getServer().createDownloader(name, config);
+        Downloader downloader = downloaderManager.createDownloader(name, config);
         if (downloader == null) {
             ctx.status(HttpStatus.BAD_REQUEST);
             ctx.json(new StdResp(false, tl(locale(ctx), Lang.DOWNLOADER_API_ADD_FAILURE), null));
